feat: icon science is a bit more fun
