+++ src/frontend/src/lib/components/icons/IconScience.svelte
@@ -1,0 +1,19 @@
+<!-- source: https://fonts.google.com/icons?selected=Material+Symbols+Outlined:science:FILL@0;wght@300;GRAD@0;opsz@24&icon.size=24&icon.color=%23000000&icon.query=scien&icon.set=Material+Symbols -->
+<script lang="ts">
+	interface Props {
+		size?: string;
+	}
+
+	let { size = '24px' }: Props = $props();
+</script>
+
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	height={size}
+	viewBox="0 -960 960 960"
+	width={size}
+	fill="currentColor"
+	><path
+		d="M200-140q-38.27 0-54.4-34.15-16.14-34.16 7.79-63.54L380-513.08V-760h-47.69q-12.75 0-21.38-8.63-8.62-8.63-8.62-21.38 0-12.76 8.62-21.37 8.63-8.62 21.38-8.62h295.38q12.75 0 21.38 8.63 8.62 8.63 8.62 21.38 0 12.76-8.62 21.37-8.63 8.62-21.38 8.62H580v246.92l226.61 275.39q23.93 29.38 7.79 63.54Q798.27-140 760-140H200Zm0-60h560L520-492v-268h-80v268L200-200Zm280-280Z"
+	/></svg
+>
