--- src/frontend/src/lib/components/icons/IconTune.svelte
@@ -1,19 +1,0 @@
-<!-- source: https://fonts.google.com/icons?selected=Material+Symbols+Outlined:discover_tune:FILL@0;wght@300;GRAD@0;opsz@24&icon.size=24&icon.color=%23000000&icon.query=tune&icon.set=Material+Symbols -->
-<script lang="ts">
-	interface Props {
-		size?: string;
-	}
-
-	let { size = '24px' }: Props = $props();
-</script>
-
-<svg
-	xmlns="http://www.w3.org/2000/svg"
-	height={size}
-	viewBox="0 -960 960 960"
-	width={size}
-	fill="currentColor"
-	><path
-		d="M530-610v-60h120v-160h60v160h120v60H530Zm120 480v-380h60v380h-60Zm-400 0v-160H130v-60h300v60H310v160h-60Zm0-320v-380h60v380h-60Z"
-	/></svg
->
